requirements.txt
PyGithub==1.59.1
GitPython==3.1.40
requests==2.31.0
python-dotenv==1.0.0
flask==2.3.3
typing-extensions==4.8.0
cryptography==41.0.7

